Added directory checks to support logging, added a hook for Patoolib and added 7z.exe as a neccesity
Added a hook for Patoolib in PyInstaller, as the program now makes use of Patoolib as it is quite extensive. Fixed the logging issue, as it needed admin perms in order to create a logging file in the program files folder, and stopped creating the necessary directories in LocalAppData, so I added a check to see if the directories needed to be made.